Move delete list button to list detail header
Adds a trash icon next to the + button in the list detail page, visible
only to the list owner. On confirm, deletes the list and redirects to /home.
Reverts delete button from home page list cards.
