+++ src/app/(app)/home/page.tsx
@@ -95,7 +95,6 @@ export default async function HomePage() {
               key={list.id}
               list={list}
               itemCount={countMap[list.id] ?? 0}
-              isOwner={list.owner_id === user.id}
             />
           ))}
         </div>
